Load custom styling before any window is shown
So the Greeter applies the custom styles too.
